fix: cast timeLeft to int when updating score on game over

+++ core/src/main/java/io/maze/screens/GameScreen.java
@@ -262,7 +262,7 @@ public class GameScreen implements Screen {
         if (timeLeft > 0) {
             timeLeft -= delta;
         } else {
-            score += timeLeft;
+            score += (int)timeLeft;
             // Transition to Game Over screen
         }
     }
